pyproject.toml
[build-system]
requires = ["setuptools >= 67.0.0"]
build-backend = "setuptools.build_meta"

[project]
name =  "conda-env-converter"
dynamic = ["version"]
description = "Build a `conda create` command from a Conda environment.yml file."
authors = [
    {name = "Rachael Esler", email = "rachael.esler@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
conda_env_converter = "conda_env_converter.__main__:main"

[tool.setuptools.dynamic]
version = {attr = "conda_env_converter.__version__"}

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
  "D",      # pydocstyle
  "E",      # pycodestyle errors
  "W",      # pycodestyle warnings
  "F",      # Pyflakes
  "I",      # isort
  "B",      # flake8-bugbear
  "ARG",    # flake8-unused-arguments
  "SIM",    # flake8-simplify
  "BLE",    # flake8-blind-except
  "FBT",    # flake8-boolean-trap
  "A",      # flake8-builtins
  "C4",     # flake8-comprehensions
  "ICN",    # flake8-import-conventions
  "T20",    # flake8-print
  "Q",      # flake8-quotes
  "RSE",    # flake8-raise
  "RET",    # flake8-return
  "SIM",    # flake8-simplify
  "PTH",    # flake8-use-pathlib
  "N",      # pep8-naming
  "PL",     # Pylint
  "RUF"     # Ruff-specific rules
]
ignore = [
  "PTH123", # Okay to use open() over Path.open()
  "T201",   # Okay to have `print` commands
  # The following rules are ignored due to conflicts with other rules
  "D203",
  "D213"
]
# Exclude package initialisation
exclude = ["src/conda_env_converter/__init__.py"]

[tool.ruff.per-file-ignores]
# No need for docstrings in __init__.py files
"__init__.py" = ["D104"]

[tool.pyright]
include = ["src"]
pythonVersion = "3.12"
